# packages/web-api/src/llm_spec_web/openapi.yaml
openapi: 3.1.0
info:
  title: llm-spec web api
  version: 0.1.0
  description: |
    Web backend for suite version management and test run orchestration.
servers:
  - url: http://localhost:8000
    description: local
paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthz
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  /api/suites:
    get:
      summary: List suites
      operationId: listSuites
      parameters:
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: endpoint
          schema:
            type: string
      responses:
        '200':
          description: suite list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuiteResponse'
    post:
      summary: Create suite (with version=1)
      operationId: createSuite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuiteCreateRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuiteResponse'

  /api/suites/{suite_id}:
    get:
      summary: Get suite
      operationId: getSuite
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      responses:
        '200':
          description: suite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuiteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update suite metadata
      operationId: updateSuite
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuiteUpdateRequest'
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuiteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete suite
      operationId: deleteSuite
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      responses:
        '204':
          description: deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /api/suites/{suite_id}/versions:
    get:
      summary: List suite versions
      operationId: listSuiteVersions
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      responses:
        '200':
          description: versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuiteVersionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create new suite version
      operationId: createSuiteVersion
      parameters:
        - $ref: '#/components/parameters/SuiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuiteVersionCreateRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuiteVersionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/suite-versions/{version_id}:
    get:
      summary: Get suite version
      operationId: getSuiteVersion
      parameters:
        - in: path
          name: version_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: suite version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuiteVersionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/provider-configs:
    get:
      summary: List provider configs
      operationId: listProviderConfigs
      responses:
        '200':
          description: provider configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProviderConfigResponse'

  /api/provider-configs/{provider}:
    get:
      summary: Get provider config
      operationId: getProviderConfig
      parameters:
        - $ref: '#/components/parameters/ProviderName'
      responses:
        '200':
          description: provider config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderConfigResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Upsert provider config
      operationId: upsertProviderConfig
      parameters:
        - $ref: '#/components/parameters/ProviderName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderConfigUpsertRequest'
      responses:
        '200':
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderConfigResponse'

  /api/runs:
    get:
      summary: List runs
      operationId: listRuns
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunJobResponse'
    post:
      summary: Create run and start in background
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreateRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunJobResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/runs/{run_id}:
    get:
      summary: Get run job
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunJobResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/runs/{run_id}/cancel:
    post:
      summary: Cancel run
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunJobResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/runs/{run_id}/events:
    get:
      summary: Poll run events
      operationId: listRunEvents
      parameters:
        - $ref: '#/components/parameters/RunId'
        - in: query
          name: after_seq
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RunEventResponse'

  /api/runs/{run_id}/events/stream:
    get:
      summary: Stream run events via SSE
      operationId: streamRunEvents
      parameters:
        - $ref: '#/components/parameters/RunId'
        - in: query
          name: after_seq
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                type: string

  /api/runs/{run_id}/result:
    get:
      summary: Get run_result.json
      operationId: getRunResult
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: run result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'

  /api/runs/{run_id}/tests:
    get:
      summary: List run test records
      operationId: listRunTests
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: tests
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true

  /api/settings/toml:
    get:
      summary: Get llm-spec.toml content
      operationId: getTomlSettings
      responses:
        '200':
          description: toml content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TomlSettingsResponse'
    put:
      summary: Update llm-spec.toml content
      operationId: updateTomlSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TomlSettingsRequest'
      responses:
        '200':
          description: updated toml content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TomlSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  parameters:
    SuiteId:
      in: path
      name: suite_id
      required: true
      schema:
        type: string
    ProviderName:
      in: path
      name: provider
      required: true
      schema:
        type: string
    RunId:
      in: path
      name: run_id
      required: true
      schema:
        type: string

  responses:
    NotFound:
      description: resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string

  schemas:
    SuiteCreateRequest:
      type: object
      required: [provider, endpoint, name, raw_json5]
      properties:
        provider:
          type: string
        endpoint:
          type: string
        name:
          type: string
        raw_json5:
          type: string
        created_by:
          type: string
          default: ui

    SuiteUpdateRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, archived]

    SuiteVersionCreateRequest:
      type: object
      required: [raw_json5]
      properties:
        raw_json5:
          type: string
        created_by:
          type: string
          default: ui

    SuiteResponse:
      type: object
      required: [id, provider, endpoint, name, status, latest_version, created_at, updated_at]
      properties:
        id:
          type: string
        provider:
          type: string
        endpoint:
          type: string
        name:
          type: string
        status:
          type: string
        latest_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SuiteVersionResponse:
      type: object
      required: [id, suite_id, version, created_by, created_at, raw_json5, parsed_json]
      properties:
        id:
          type: string
        suite_id:
          type: string
        version:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        raw_json5:
          type: string
        parsed_json:
          type: object
          additionalProperties: true

    ProviderConfigUpsertRequest:
      type: object
      required: [base_url, timeout, api_key]
      properties:
        base_url:
          type: string
        timeout:
          type: number
        api_key:
          type: string
        extra_config:
          type: object
          additionalProperties: true

    ProviderConfigResponse:
      type: object
      required: [provider, base_url, timeout, extra_config, updated_at]
      properties:
        provider:
          type: string
        base_url:
          type: string
        timeout:
          type: number
        extra_config:
          type: object
          additionalProperties: true
        updated_at:
          type: string
          format: date-time

    RunCreateRequest:
      type: object
      required: [suite_version_id]
      properties:
        suite_version_id:
          type: string
        mode:
          type: string
          enum: [real, mock]
          nullable: true
        selected_tests:
          type: array
          items:
            type: string

    RunJobResponse:
      type: object
      required:
        [id, status, mode, provider, endpoint, progress_total, progress_done, progress_passed, progress_failed]
      properties:
        id:
          type: string
        status:
          type: string
        mode:
          type: string
        provider:
          type: string
        endpoint:
          type: string
        suite_version_id:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        progress_total:
          type: integer
        progress_done:
          type: integer
        progress_passed:
          type: integer
        progress_failed:
          type: integer
        error_message:
          type: string
          nullable: true

    RunEventResponse:
      type: object
      required: [id, run_id, seq, event_type, payload, created_at]
      properties:
        id:
          type: integer
        run_id:
          type: string
        seq:
          type: integer
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    TomlSettingsRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string

    TomlSettingsResponse:
      type: object
      required: [path, content, exists]
      properties:
        path:
          type: string
        content:
          type: string
        exists:
          type: boolean